requirements.txt
flask
psycopg2-binary
apscheduler
selenium
webdriver-manager
requests
beautifulsoup4
